+++ .ci/scripts/native_test_win32.sh
@@ -3,5 +3,5 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-dart ./script/tool/bin/flutter_plugin_tools.dart native-test --windows \
+dart pub global run flutter_plugin_tools native-test --windows \
    --no-integration --packages-for-branch --log-timing