+++ packages/in_app_purchase/in_app_purchase_storekit/lib/src/types/app_store_purchase_param.dart
@@ -10,15 +10,12 @@ import '../../store_kit_wrappers.dart';
 class AppStorePurchaseParam extends PurchaseParam {
   /// Creates a new [AppStorePurchaseParam] object with the given data.
   AppStorePurchaseParam({
-    required ProductDetails productDetails,
-    String? applicationUserName,
+    required super.productDetails,
+    super.applicationUserName,
     this.quantity = 1,
     this.simulatesAskToBuyInSandbox = false,
     this.discount,
-  }) : super(
-          productDetails: productDetails,
-          applicationUserName: applicationUserName,
-        );
+  });
 
   /// Set it to `true` to produce an "ask to buy" flow for this payment in the
   /// sandbox.