+++ packages/in_app_purchase/in_app_purchase_storekit/test/fakes/fake_storekit_platform.dart
@@ -14,7 +14,9 @@ import '../store_kit_wrappers/sk_test_stub_objects.dart';
 
 class FakeStoreKitPlatform {
   FakeStoreKitPlatform() {
-    channel.setMockMethodCallHandler(onMethodCall);
+    _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+        .defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, onMethodCall);
   }
 
   // pre-configured store information
@@ -235,3 +237,9 @@ class FakeStoreKitPlatform {
     return (call.arguments as Map<Object?, Object?>).cast<String, Object?>();
   }
 }
+
+/// This allows a value of type T or T? to be treated as a value of type T?.
+///
+/// We use this so that APIs that have become non-nullable can still be used
+/// with `!` and `?` on the stable branch.
+T? _ambiguate<T>(T? value) => value;