+++ packages/local_auth/local_auth_platform_interface/test/default_method_channel_platform_test.dart
@@ -29,7 +29,9 @@ void main() {
   });
 
   test('getAvailableBiometrics', () async {
-    channel.setMockMethodCallHandler((MethodCall methodCall) {
+    _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+        .defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, (MethodCall methodCall) {
       log.add(methodCall);
       return Future<dynamic>.value(<String>[]);
     });
@@ -49,7 +51,9 @@ void main() {
     // existing unendorsed implementations, used 'undefined' as a special
     // return value from `getAvailableBiometrics` to indicate that nothing was
     // enrolled, but that the hardware does support biometrics.
-    channel.setMockMethodCallHandler((MethodCall methodCall) {
+    _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+        .defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, (MethodCall methodCall) {
       log.add(methodCall);
       return Future<dynamic>.value(<String>['undefined']);
     });
@@ -68,7 +72,9 @@ void main() {
 
   group('Boolean returning methods', () {
     setUp(() {
-      channel.setMockMethodCallHandler((MethodCall methodCall) {
+      _ambiguate(TestDefaultBinaryMessengerBinding.instance)!
+          .defaultBinaryMessenger
+          .setMockMethodCallHandler(channel, (MethodCall methodCall) {
         log.add(methodCall);
         return Future<dynamic>.value(true);
       });
@@ -198,3 +204,9 @@ void main() {
     });
   });
 }
+
+/// This allows a value of type T or T? to be treated as a value of type T?.
+///
+/// We use this so that APIs that have become non-nullable can still be used
+/// with `!` and `?` on the stable branch.
+T? _ambiguate<T>(T? value) => value;