+++ packages/webview_flutter/webview_flutter/lib/src/navigation_delegate.dart
@@ -12,6 +12,28 @@ import 'webview_controller.dart';
 /// the progress of navigation requests.
 ///
 /// See [WebViewController.setNavigationDelegate].
+///
+/// ## Platform-Specific Features
+/// This class contains an underlying implementation provided by the current
+/// platform. Once a platform implementation is imported, the examples below
+/// can be followed to use features provided by a platform's implementation.
+///
+/// {@macro webview_flutter.NavigationDelegate.fromPlatformCreationParams}
+///
+/// Below is an example of accessing the platform-specific implementation for
+/// iOS and Android:
+///
+/// ```dart
+/// final NavigationDelegate navigationDelegate = NavigationDelegate();
+///
+/// if (WebViewPlatform.instance is WebKitWebViewPlatform) {
+///   final WebKitNavigationDelegate webKitDelegate =
+///       navigationDelegate.platform as WebKitNavigationDelegate;
+/// } else if (WebViewPlatform.instance is AndroidWebViewPlatform) {
+///   final AndroidNavigationDelegate androidDelegate =
+///       navigationDelegate.platform as AndroidNavigationDelegate;
+/// }
+/// ```
 class NavigationDelegate {
   /// Constructs a [NavigationDelegate].
   NavigationDelegate({
@@ -32,6 +54,33 @@ class NavigationDelegate {
 
   /// Constructs a [NavigationDelegate] from creation params for a specific
   /// platform.
+  ///
+  /// {@template webview_flutter.NavigationDelegate.fromPlatformCreationParams}
+  /// Below is an example of setting platform-specific creation parameters for
+  /// iOS and Android:
+  ///
+  /// ```dart
+  /// PlatformNavigationDelegateCreationParams params =
+  ///     const PlatformNavigationDelegateCreationParams();
+  ///
+  /// if (WebViewPlatform.instance is WebKitWebViewPlatform) {
+  ///   params = WebKitNavigationDelegateCreationParams
+  ///       .fromPlatformNavigationDelegateCreationParams(
+  ///     params,
+  ///   );
+  /// } else if (WebViewPlatform.instance is AndroidWebViewPlatform) {
+  ///   params = AndroidNavigationDelegateCreationParams
+  ///       .fromPlatformNavigationDelegateCreationParams(
+  ///     params,
+  ///   );
+  /// }
+  ///
+  /// final NavigationDelegate navigationDelegate =
+  ///     NavigationDelegate.fromPlatformCreationParams(
+  ///   params,
+  /// );
+  /// ```
+  /// {@endtemplate}
   NavigationDelegate.fromPlatformCreationParams(
     PlatformNavigationDelegateCreationParams params, {
     FutureOr<NavigationDecision> Function(NavigationRequest request)?