+++ packages/webview_flutter/webview_flutter_wkwebview/ios/Classes/FLTWebViewFlutterPlugin.h
@@ -3,6 +3,11 @@
 // found in the LICENSE file.
 
 #import <Flutter/Flutter.h>
+#import <WebKit/WebKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
 
 @interface FLTWebViewFlutterPlugin : NSObject <FlutterPlugin>
 @end
+
+NS_ASSUME_NONNULL_END