+++ packages/in_app_purchase/in_app_purchase/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.1.4
+
+* Updates iOS minimum version in README.
+
+## 3.1.3
+
+* Ignores a lint in the example app for backwards compatibility.
+
 ## 3.1.2
 
 * Updates example code for `use_build_context_synchronously` lint.