+++ packages/shared_preferences/shared_preferences_foundation/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.1.3
 
+* Uses the new `sharedDarwinSource` flag when available.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.1.2