+++ packages/webview_flutter/webview_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.0.4
+
+* Adds examples of accessing platform-specific features for each class.
+
 ## 4.0.3
 
 * Updates example code for `use_build_context_synchronously` lint.